do not run things on non-ubuntu
Former-commit-id: 195eb03dac8c83563809c99b44eac072e5c04bba
Former-commit-id: 8c30380b80c5d00f76631822fedf748b29a9842f
Former-commit-id: 00a3719b3c80f603743fcc54c4bd6a1e210e72ad

+++ source/.chezmoiscripts/run_once_set-splash.sh
@@ -1,25 +1,34 @@
 #!/bin/bash
 # Install vortex-ubuntu-plymouth-theme splash screen
 
-tempdir=$(mktemp -d) && (
-  # Stop on error and remove temporary dir
-  set -e
-  trap 'rm -rf "$tempdir"' EXIT
+set -e
 
-  # Clone and cd into repo
-  git clone 'https://github.com/emanuele-scarsella/vortex-ubuntu-plymouth-theme' "$tempdir" --depth 1
-  cd "$tempdir"
+function log() { printf '%s\n' "$@"; }
+function success() { printf "$(tput setaf 2)%s$(tput sgr0)\n" "Success!"; }
 
-  # Create conf.d to show the splash on boot
-  if [[ -f /etc/initramfs-tools/conf.d/splash ]]; then
-    printf '%s\n' 'Overriding /etc/initramfs-tools/conf.d/splash'
-    sleep 3 # Allow user to halt
-  fi
-  printf 'FRAMEBUFFER=y' | sudo tee /etc/initramfs-tools/conf.d/splash >/dev/null
-  # Runing update is not necessary because the install script will do it
+log "Installing vortex-ubuntu-plymouth-theme splash screen"
+os="$(lsb_release -is)"
+if [ "$os" != "Ubuntu" ]; then
+  log "This script is only available for Ubuntu" >&2
+  exit 0
+fi
 
-  # Run install script
-  sudo chmod +x "$tempdir/install"
-  sudo "$tempdir/install"
+tempdir=$(mktemp -d)
+# Stop on error and remove temporary dir
+trap 'rm -rf "$tempdir"' EXIT
 
-)
+# Clone and cd into repo
+git clone 'https://github.com/emanuele-scarsella/vortex-ubuntu-plymouth-theme' "$tempdir" --depth 1
+cd "$tempdir"
+
+# Create conf.d to show the splash on boot
+if [[ -f /etc/initramfs-tools/conf.d/splash ]]; then
+  printf '%s\n' 'Overriding /etc/initramfs-tools/conf.d/splash'
+  sleep 3 # Allow user to halt
+fi
+printf 'FRAMEBUFFER=y' | sudo tee /etc/initramfs-tools/conf.d/splash >/dev/null
+# Runing update is not necessary because the install script will do it
+
+# Run install script
+sudo chmod +x "$tempdir/install"
+sudo "$tempdir/install"
